refactor duplicated code
create common.js for common functions
