+++ lib/adapter.js
@@ -104,7 +104,10 @@ karma.start = () => {
                 workerPort = worker.port;
         }
         else {
-                throw new Error('[karma-mocha-webworker] Unexpected Worker Class ' + workerType);
+                throw new Error(
+                        '[karma-mocha-webworker] Invalid karma configuration value for "client.mochaWebWorker.worker". ' +
+                        'Expected "Worker" or "SharedWorker" instead of "' + workerType + '"'
+                );
         }
 
         const rpc = new JSONBird({
